+++ packages/downloading-helpers/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @alwaysmeticulous/downloading-helpers
 
+## 2.337.1
+
+### Patch Changes
+
+- Updated dependencies [[`376e029`](https://github.com/alwaysmeticulous/meticulous/commit/376e02973a67a487dd12e6e02c74897016f7b784), [`1320f6a`](https://github.com/alwaysmeticulous/meticulous/commit/1320f6a4a5a3d8c3287a4d8dde66e5d4166fedd9), [`1b49d0e`](https://github.com/alwaysmeticulous/meticulous/commit/1b49d0e737ac543546011131973c213b36fa6cdf)]:
+  - @alwaysmeticulous/client@2.337.1
+
 ## 2.337.0
 
 ### Patch Changes